By issuing a PATCH request directly to the user path, verify that the admin attribute
isn’t editable through the web.

if we simply passed an initialization hash in from an arbitrary web request,
a malicious user could send a PATCH request as follows:

patch /users/17?admin=1

This request would make user 17 an admin, which would be a potentially serious security breach.
Because of this danger, it is essential that we only update attributes that are safe to edit
through the web. As noted in Section 7.3.2, this is accomplished using strong parameters by
calling require and permit on the params hash:

  private

patch /users/17?admin=1

Because of this danger, it is essential that we only update attributes that are safe to edit through the web.
As noted in Section 7.3.2, this is accomplished using strong parameters by calling require and permit
on the params hash:

private
def user_params
  params.require(:user).permit(:name, :email, :password, :password_confirmation)
end