Optimize the feed method to fix the scaling probem
Previous doesn’t scale well when the number of microposts in the feed is large, as would
likely happen if a user were following, say, 5000 other users.

The problem with the code in Section 14.3.2 is that following_ids pulls all
the followed users’ ids into memory, and creates an array the full length
of the followed users array. Since the condition in Listing 14.44
actually just checks inclusion in a set, there must be a more efficient
way to do this, and indeed SQL is optimized for just such set operations.

The solution involves pushing the finding of followed user ids into the
database using a subselect.

* Using key-value pairs in the feed’s where method